requirements.txt
gunicorn==21.2.0
aiosmtplib==4.0.0
alembic==1.15.2
amqp==5.3.1
billiard==4.2.1
blinker==1.9.0
cachelib==0.13.0
celery==5.5.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Deprecated==1.2.18
Flask==3.1.0
Flask-Limiter==3.11.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
kombu==5.5.2
limits==4.6
Mako==1.3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
msgspec==0.19.0
mysql-connector-python==9.2.0
mysqlclient==2.2.7
numpy==2.2.4
ordered-set==4.1.0
packaging==24.2
pandas==2.2.3
prompt_toolkit==3.0.50
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
redis==5.2.1
rich==13.9.4
six==1.17.0
SQLAlchemy==2.0.40
typing_extensions==4.13.1
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
wrapt==1.17.2
